NF: move TagsArrayAdapter to its own file
